Provide test functionality in traits
Fix #202.

separate http client from trait

Make property and methods private

Add traits to docs

Reduce levels of inheritance
